refactor: remove `linux` subfolder
Remove the separate `linux` subfolder of the
epoll & eventfd bindings in `lib/src/` and `lib/src/gnu`.
