+++ lib/src/gnu/gnu.dart
@@ -5,8 +5,8 @@ import 'errno.dart';
 import 'fcntl.dart';
 import 'ffigen.dart' as ffi;
 import 'glob.dart';
-import 'linux/epoll.dart';
-import 'linux/eventfd.dart';
+import 'epoll.dart';
+import 'eventfd.dart';
 import 'macros.g.dart';
 import 'mman.dart';
 import 'pwd.dart';
